Coins and a "Bus"
Added and animated gold and silver coins that are destroyed when the
player collides with them. Would like to see them spawn from dead
enemies and affect a player score eventually. (This would add some
competitive aspects to the game.)

Also added a "bus" (aka a couple blocks and a big collider) behind the
player spawn point to prevent players from running off the map
backwards.